Minor update regarding trtllm fp8 kvcache functionality issue
https://github.com/sgl-project/sglang/issues/9704#issuecomment-3231432367

Accuracy
Accuracy: 0.964
Invalid: 0.000
Latency: 106.218 s
Output throughput: 1348.309 token/s

Commands:
```
SGL_ENABLE_JIT_DEEPGEMM=0  python3 -m sglang.launch_server --tokenizer-path deepseek-ai/DeepSeek-R1-0528 --trust-remote-code --enable-dp-attention --disable-radix-cache  --mem-fraction-static 0.89 --model-path=deepseek-ai/DeepSeek-R1-0528 --host 0.0.0.0 --port 8000 --tensor-parallel-size=8 --data-parallel-size=8  --attention-backend flashinfer
```
```
python3 benchmark/gsm8k/bench_sglang.py --num-shots 8 --num-questions 1316 --parallel 1316 --port 8000
```

+++ python/sglang/srt/models/deepseek_v2.py
@@ -1669,7 +1669,7 @@ class DeepseekV2AttentionMLA(nn.Module):
             # Fetch latent cache from memory pool with precomputed chunked kv indices
             latent_cache_buf = forward_batch.token_to_kv_pool.get_key_buffer(
                 self.attn_mha.layer_id
-            )
+            ).to(q.dtype)
             latent_cache = latent_cache_buf[
                 forward_batch.prefix_chunk_kv_indices[i]
             ].contiguous()
